requirements.txt
playwright==1.53.0
python-dotenv==1.1.1
requests==2.32.3
Flask==3.1.1
